Fix dependabot-update script
`--remote` is not a recognized flag for Git. I swear it used to be.
Maybe it was removed in a recent version?

The script also failed if you have unstaged changes. Stashing them fixes
this. As a future enhancement the script could put `git stash pop` in a
trap function so it happens even if the script ends abnormally.