Track last refresh time
This allows us to more accurately enforce the min/max refresh interval
with (I think) more readable code. The code is more complicated than the
prior commit, but that's because the new structure made several bugs
more apparent.

It would be simpler still if we just had one refresh interval, but I
kind of like the idea of keeping the app only minimally up to date when
the user is not interacting with it, but allowing more frequent updates
when the user keep checking.